third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpecularLightingElement-dom-specularExponent-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'specularExponent' attribute of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS specularElement.getAttribute('specularExponent') is "4"
PASS specularElement.getAttribute('specularExponent') is "1"
PASS successfullyParsed is true

TEST COMPLETE
